Added README.md for Clarity
#Hello world!#

I am currently trying to build a thermostat page in collaboration with others as a means of learning and implementing the interactive functionality of javascript. 

For the AJAX request functionality, I have currently decided to substitute the API key for a junk variable as I feel that here may be a good place to implement an environmental variable in future as practice for handling more sensitive information in the future.
I also want to further improve the styling of the interface via some external CSS.

Watch this space!